introduce-lora-controller-move: Hand LoRA load/unload over to LoraController
Pure physical move per MECH_COMMIT_SPLIT. Cut the 6 @staticmethod lora
methods (4 from TokenizerControlMixin: load_lora_adapter,
load_lora_adapter_from_tensors, unload_lora_adapter,
_unload_lora_adapter_locked; 2 from TokenizerManager:
_validate_and_resolve_lora, _resolve_lora_path); paste into LoraController
(drop @staticmethod, replace ``self: "LoraController"`` → plain ``self``).
Caller prefix replacement:
``TokenizerManager.<method>(self.lora_controller, ...)`` →
``self.lora_controller.<method>(...)``; ditto entrypoints.

Refactor chain ID: introduce-lora-controller-move
