+++ python/sglang/srt/managers/tokenizer_manager.py
@@ -53,7 +53,6 @@ from sglang.srt.managers.io_struct import (
     FreezeGCReq,
     GenerateReqInput,
     HealthCheckOutput,
-    LoadLoRAAdapterReqInput,
     OpenSessionReqOutput,
     PauseGenerationReqInput,
     TokenizedEmbeddingReqInput,
@@ -441,7 +440,7 @@ class TokenizerManager(TokenizerControlMixin):
             await self.is_pause_cond.wait_for(lambda: not self.is_pause)
 
         async with self.model_update_lock.reader_lock:
-            await TokenizerManager._validate_and_resolve_lora(self.lora_controller, obj)
+            await self.lora_controller._validate_and_resolve_lora(obj)
 
             # Tokenize the request and send it to the scheduler
             if obj.is_single:
@@ -1252,88 +1251,6 @@ class TokenizerManager(TokenizerControlMixin):
 
     def update_active_ranks(self, ranks: ActiveRanksOutput):
         self.send_to_scheduler.send_pyobj(ranks)
-
-    @staticmethod
-    async def _validate_and_resolve_lora(
-        self: "LoraController",
-        obj: Union[GenerateReqInput, EmbeddingReqInput],
-    ) -> None:
-        if not obj.lora_path:
-            return
-
-        if not self.server_args.enable_lora:
-            first_adapter = (
-                obj.lora_path
-                if isinstance(obj.lora_path, str)
-                else next((a for a in obj.lora_path if a), None)
-            )
-
-            raise ValueError(
-                f"LoRA adapter '{first_adapter}' was requested, but LoRA is not enabled. "
-                "Please launch the server with --enable-lora flag and preload adapters "
-                "using --lora-paths or /load_lora_adapter endpoint."
-            )
-
-        await TokenizerManager._resolve_lora_path(self, obj)
-
-    @staticmethod
-    async def _resolve_lora_path(
-        self: "LoraController",
-        obj: Union[GenerateReqInput, EmbeddingReqInput],
-    ):
-        if isinstance(obj.lora_path, str):
-            unique_lora_paths = set([obj.lora_path])
-        else:
-            unique_lora_paths = set(obj.lora_path)
-
-        if (
-            self.server_args.max_loaded_loras is not None
-            and len(unique_lora_paths) > self.server_args.max_loaded_loras
-        ):
-            raise ValueError(
-                f"Received request with {len(unique_lora_paths)} unique loras requested "
-                f"but max loaded loras is {self.server_args.max_loaded_loras}"
-            )
-
-        # Reload all existing LoRA adapters that have been dynamically unloaded
-        unregistered_loras = await self.lora_registry.get_unregistered_loras(
-            unique_lora_paths
-        )
-        for lora_path in unregistered_loras:
-            if lora_path is None:
-                continue
-
-            if lora_path not in self.lora_ref_cache:
-                raise ValueError(
-                    f"Got LoRA adapter that has never been loaded: {lora_path}\n"
-                    f"All loaded adapters: {self.lora_ref_cache.keys()}."
-                )
-
-            logger.info(f"Reloading evicted adapter: {lora_path}")
-            new_lora_ref = self.lora_ref_cache[lora_path]
-            load_result = await TokenizerManager.load_lora_adapter(
-                self,
-                LoadLoRAAdapterReqInput(
-                    lora_name=new_lora_ref.lora_name,
-                    lora_path=new_lora_ref.lora_path,
-                    pinned=new_lora_ref.pinned,
-                ),
-            )
-            if (
-                not load_result.success
-                and "already loaded" not in load_result.error_message
-            ):
-                raise ValueError(
-                    f"Failed to implicitly load LoRA adapter {lora_path}: {load_result.error_message}"
-                )
-
-        # Look up the LoRA ID from the registry and start tracking ongoing LoRA requests.
-        obj.lora_id = await self.lora_registry.acquire(obj.lora_path)
-        # Propagate lora_id to any sub-objects already cached by __getitem__.
-        for i, sub_obj in obj.__dict__.get("_sub_obj_cache", {}).items():
-            sub_obj.lora_id = (
-                obj.lora_id[i] if isinstance(obj.lora_id, list) else obj.lora_id
-            )
 
     def _set_default_priority(self, obj: Union[GenerateReqInput, EmbeddingReqInput]):
         """Set the default priority value."""
